# src/main/resources/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "alojamiento quindio",
    "description": "API REST para la gestión completa",
    "contact": {
      "name": "Equipo de Desarrollo",
      "url": "https://tiendavirtual.com",
      "email": "dev@tiendavirtual.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de Desarrollo"
    },
    {
      "url": "https://api-test.tiendavirtual.com",
      "description": "Servidor de Pruebas"
    },
    {
      "url": "https://api.tiendavirtual.com",
      "description": "Servidor de Producción"
    }
  ],
...
  "paths": {
...
    "/api/v1/accommodations/{id}": {
      "get": {
        "tags": [
          "Accommodations"
        ],
        "summary": "Obtener detalles de alojamiento",
        "description": "Incluye mapa y comentarios promedio.",
        "operationId": "getAccommodationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del alojamiento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del alojamiento",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Accommodations"
        ],
        "summary": "Actualizar alojamiento",
        "description": "Solo para host propietario. Soft update.",
        "operationId": "updateAccommodation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del alojamiento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "403": {
            "description": "No propietario",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Accommodations"
        ],
        "summary": "Soft delete de alojamiento",
        "description": "Marca como deleted; solo para host propietario.",
        "operationId": "deleteAccommodation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del alojamiento",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado lógicamente"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No propietario"
          },
          "404": {
            "description": "No encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/accommodations": {
      "get": {
        "tags": [
          "Accommodations"
        ],
        "summary": "Buscar alojamientos disponibles",
        "description": "Consulta por ciudad, fechas, precio; integra Mapbox para ubicaciones. Filtros opcionales.",
        "operationId": "searchAccommodations",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "description": "Ciudad para filtrar",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Madrid"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Fecha inicio disponibilidad",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2025-10-01"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Fecha fin disponibilidad",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2025-10-10"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Precio máximo por noche",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 100
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Tamaño de página",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de alojamientos disponibles",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccommodationDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (e.g., fechas erróneas)",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccommodationDTO"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccommodationDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accommodations"
        ],
        "summary": "Crear nuevo alojamiento",
        "description": "Solo para hosts. Incluye hasta 10 imágenes (URLs de servicio externo).",
        "operationId": "createAccommodation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Alojamiento creado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida (e.g., imágenes \u003E10)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          },
          "403": {
            "description": "Rol no autorizado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccommodationUpdateDTO": {
        "description": "Datos actualizados del alojamiento"
      },
      "AccommodationDTO": {

      },
...
      "AccommodationCreateDTO": {
        "description": "Datos del alojamiento a crear"
      },
...
    }
  }
}